intaller: moves config funcs into respectively named files
