+++ pkg/install/app/app.go
@@ -6,6 +6,13 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
+const (
+	webhookTLSDirPath     = "/var/run/secrets/puppet/relay/webhook-tls"
+	jwtSigningKeyDirPath  = "/var/run/secrets/puppet/relay/jwt"
+	jwtSigningKeyPath     = "/var/run/secrets/puppet/relay/jwt/private-key.pem"
+	metadataAPITLSDirPath = "/var/run/secrets/puppet/relay/tls"
+)
+
 var DependencyManager = ownerext.NewManager("installer.relay.sh/dependency-of")
 
 // TODO share this with pkg/operator/app
